+++ lib/osm/evening.rb
@@ -6,7 +6,7 @@ module Osm
     attr_reader :start_time, :end_time
     # @!attribute [rw] evening_id
     #   @return [Fixnum] the id of the evening
-    # @!attribute [rw] sectionid
+    # @!attribute [rw] section_id
     #   @return [Fixnum] the section the evening belongs to
     # @!attribute [rw] title
     #   @return [String] the title of the evening
@@ -29,29 +29,53 @@ module Osm
     # @!attribute [rw] end_time
     #   @return [String] the end time (hh:mm)
 
-    # Initialize a new ProgrammeItem using the hash returned by the API call
-    # @param data the hash of data for the object returned by the API
-    # @param activities an array of hashes to generate the list of ProgrammeActivity objects
-    def initialize(data, activities)
-      @evening_id = Osm::to_i_or_nil(data['eveningid'])
-      @section_id = Osm::to_i_or_nil(data['sectionid'])
-      @title = data['title'] || 'Unnamed meeting'
-      @notes_for_parents = data['notesforparents'] || ''
-      @games = data['games'] || ''
-      @pre_notes = data['prenotes'] || ''
-      @post_notes = data['postnotes'] || ''
-      @leaders = data['leaders'] || ''
-      @start_time = data['starttime'].nil? ? nil : data['starttime'][0..4]
-      @end_time = data['endtime'].nil? ? nil : data['endtime'][0..4]
-      @meeting_date = Osm::parse_date(data['meetingdate'])
+    # Initialize a new Evening
+    # @param [Hash] attributes the hash of attributes (see attributes for descriptions, use Symbol of attribute name as the key)
+    def initialize(attributes={})
+      [:evening_id, :section_id].each do |attribute|
+        raise ArgumentError, ":#{attribute} must be nil or a Fixnum > 0" unless attributes[attribute].nil? || (attributes[attribute].is_a?(Fixnum) && attributes[attribute] > 0)
+      end
+      [:title, :notes_for_parents, :games, :pre_notes, :post_notes, :leaders].each do |attribute|
+        raise ArgumentError, ":#{attribute} must be nil or a String" unless (attributes[attribute].nil? || attributes[attribute].is_a?(String))
+      end
+      raise ArgumentError, ':meeting_date must be a Date' unless attributes[:meeting_date].is_a?(Date)
+      raise ArgumentError, ':activities must be nil or an Array of Osm::Evening::Activity' unless (attributes[:activities].nil? || Osm::is_array_of?(attributes[:activities], Osm::Evening::Activity))
 
-      @activities = Array.new
+      attributes.each { |k,v| send("#{k}=", v) }
+
+      @activities ||= []
+      @title ||= 'Unnamed meeting'
+      [:notes_for_parents, :games, :pre_notes, :post_notes, :leaders].each do |attribute|
+        instance_variable_set("@#{attribute}", '') if instance_variable_get("@#{attribute}").nil?
+      end
+    end
+
+
+    # Initialize a new Evening from api data
+    # @param [Hash] data the hash of data provided by the API
+    # @param activities an array of hashes to generate the list of ProgrammeActivity objects
+    def self.from_api(data, activities)
+      attributes = {}
+      attributes[:evening_id] = Osm::to_i_or_nil(data['eveningid'])
+      attributes[:section_id] = Osm::to_i_or_nil(data['sectionid'])
+      attributes[:title] = data['title'] || 'Unnamed meeting'
+      attributes[:notes_for_parents] = data['notesforparents'] || ''
+      attributes[:games] = data['games'] || ''
+      attributes[:pre_notes] = data['prenotes'] || ''
+      attributes[:post_notes] = data['postnotes'] || ''
+      attributes[:leaders] = data['leaders'] || ''
+      attributes[:start_time] = data['starttime'].nil? ? nil : data['starttime'][0..4]
+      attributes[:end_time] = data['endtime'].nil? ? nil : data['endtime'][0..4]
+      attributes[:meeting_date] = Osm::parse_date(data['meetingdate'])
+
+      attributes[:activities] = Array.new
       unless activities.nil?
         activities.each do |item|
-          @activities.push Activity.new(item)
+          attributes[:activities].push Evening::Activity.from_api(item)
         end
       end
-      @activities.freeze
+
+      new(attributes)
     end
 
     # Custom setters for times
@@ -71,7 +95,7 @@ module Osm
       {
         'eveningid' => evening_id,
         'sectionid' => section_id,
-        'meetingdate' => meeting_date.try(:strftime, '%Y-%m-%d'),
+        'meetingdate' => meeting_date.strftime('%Y-%m-%d'),
         'starttime' => start_time,
         'endtime' => end_time,
         'title' => title,
@@ -111,12 +135,25 @@ module Osm
       # @!attribute [r] notes
       #   @return [String] notes relevant to doing this activity on this evening
   
-      # Initialize a new Activity using the hash returned by the API call
-      # @param data the hash of data for the object returned by the API
-      def initialize(data)
-        @activity_id = Osm::to_i_or_nil(data['activityid'])
-        @title = data['title']
-        @notes = data['notes']
+      # Initialize a new Evening::Activity
+      # @param [Hash] attributes the hash of attributes (see attributes for descriptions, use Symbol of attribute name as the key)
+      def initialize(attributes={})
+        raise ArgumentError, ':activity_id must be a Fixnum > 0' unless (attributes[:activity_id].is_a?(Fixnum) && attributes[:activity_id] > 0)
+        raise ArgumentError, ':title must be nil or a String' unless (attributes[:title].nil? || attributes[:title].is_a?(String))
+        raise ArgumentError, ':notes must be nil or a String' unless (attributes[:notes].nil? || attributes[:notes].is_a?(String))
+  
+        attributes.each { |k,v| instance_variable_set("@#{k}", v) }
+      end
+
+
+      # Initialize a new Evening::Activity from api data
+      # @param [Hash] data the hash of data provided by the API
+      def self.from_api(data)
+        new({
+          :activity_id => Osm::to_i_or_nil(data['activityid']),
+          :title => data['title'],
+          :notes => data['notes'],
+        })
       end
 
     end