+++ spec/osm/evening_spec.rb
@@ -4,7 +4,23 @@ require 'date'
 
 describe "Evening" do
 
-  it "Create" do
+  before :each do
+    @attributes = {
+      :evening_id => 1,
+      :section_id => 2,
+      :title => 'Evening Name',
+      :notes_for_parents => 'Notes for parents',
+      :games => 'Games',
+      :pre_notes => 'Before',
+      :post_notes => 'After',
+      :leaders => 'Leaders',
+      :start_time => '19:00',
+      :end_time => '21:00',
+      :meeting_date => Date.new(2000, 01, 02),
+    }
+  end
+
+  it "Create from API data" do
     data = {
       'eveningid' => 1,
       'sectionid' => 2,
@@ -24,7 +40,7 @@ describe "Evening" do
       'title' => 'Activity Name',
       'notes' => 'Notes',
     }]
-    e = Osm::Evening.new(data, activities)
+    e = Osm::Evening.from_api(data, activities)
 
     e.evening_id.should == 1
     e.section_id.should == 2
@@ -38,7 +54,6 @@ describe "Evening" do
     e.end_time.should == '21:00'
     e.meeting_date.should == Date.new(2000, 1, 2)
 
-    e.activities.frozen?.should be_true
     ea = e.activities[0]
     ea.activity_id.should == 2
     ea.title.should == 'Activity Name'
@@ -47,7 +62,7 @@ describe "Evening" do
 
 
   it "Raises exceptions when trying to set invalid times" do
-    e = Osm::Evening.new({}, [])
+    e = Osm::Evening.new(@attributes)
     
     expect{ e.start_time = 'abcde' }.to raise_error(ArgumentError)
     expect{ e.start_time = '24:00' }.to raise_error(ArgumentError)
@@ -64,26 +79,16 @@ describe "Evening" do
 
 
   it "Creates the data for saving through the API" do
-    data = {
-      'eveningid' => 1,
-      'sectionid' => 2,
-      'title' => 'Evening Name',
-      'notesforparents' => 'Notes for parents',
-      'games' => 'Games',
-      'prenotes' => 'Before',
-      'postnotes' => 'After',
-      'leaders' => 'Leaders',
-      'starttime' => '19:00',
-      'endtime' => '21:00',
-      'meetingdate' => '2000-01-02',
-    }
-    activities = [{
-      'eveningid' => 3,
-      'activityid' => 4,
-      'title' => 'Activity Name',
-      'notes' => 'Notes',
-    }]
-    e = Osm::Evening.new(data, activities)
+    data = @attributes.merge(
+      :activities => [ Osm::Evening::Activity.new(
+        :activity_id => 4,
+        :title => 'Activity Name',
+        :notes => 'Notes',
+      ) ]
+    )
+
+    e = Osm::Evening.new(data)
+
     e.to_api.should == {
       'eveningid' => 1,
       'sectionid' => 2,